大数据系统大作业/requirements.txt
# 爬虫相关
scrapy>=2.11.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
lxml>=4.9.0

# 数据处理
pandas>=2.1.0
numpy>=1.24.0

# HBase客户端
happybase>=1.2.0
thrift>=0.19.0

# 搜索引擎和文本处理
jieba>=0.42.1
scikit-learn>=1.3.0

# Web框架
flask>=3.0.0
flask-cors>=4.0.0

# 工具库
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0.1

# 文件处理
python-docx>=1.1.0
openpyxl>=3.1.0
PyPDF2>=3.0.0

# 其他工具
tqdm>=4.66.0
python-dotenv>=1.0.0


